Fix PreShooter
